Split out Frame{Flags,Type} and ErrorCode out from Frame.h
Trying to reduce the surface area of Frame.h and internal/Common.h.

Cut out some duplicate definitions of StreamId as well.  It'll stay defined in
internal/Common.h for now.
