+++ rsocket/Payload.h
@@ -5,23 +5,19 @@
 #include <folly/io/IOBuf.h>
 #include <memory>
 #include <string>
-#include "rsocket/internal/Common.h"
+
+#include "rsocket/framing/FrameFlags.h"
 
 namespace rsocket {
-
-enum class FrameFlags : uint16_t;
 
 /// The type of a read-only view on a binary buffer.
 /// MUST manage the lifetime of the underlying buffer.
 struct Payload {
   Payload() = default;
-  explicit Payload(
-      std::unique_ptr<folly::IOBuf> _data,
-      std::unique_ptr<folly::IOBuf> _metadata =
-          std::unique_ptr<folly::IOBuf>());
-  explicit Payload(
-      const std::string& data,
-      const std::string& metadata = std::string());
+  Payload(
+      std::unique_ptr<folly::IOBuf> data,
+      std::unique_ptr<folly::IOBuf> metadata = std::unique_ptr<folly::IOBuf>());
+  Payload(const std::string& data, const std::string& metadata = std::string());
 
   explicit operator bool() const {
     return data != nullptr || metadata != nullptr;
@@ -41,4 +37,4 @@ struct Payload {
 };
 
 std::ostream& operator<<(std::ostream& os, const Payload& payload);
-} // reactivesocket
+}
