+++ rsocket/statemachine/StreamState.h
@@ -14,7 +14,6 @@ namespace rsocket {
 class RSocketStateMachine;
 class RSocketStats;
 class StreamStateMachineBase;
-using StreamId = uint32_t;
 
 class StreamState {
  public:
